+++ Sources/Sentry/SentryBaggage.m
@@ -14,9 +14,8 @@
                       publicKey:(NSString *)publicKey
                     releaseName:(nullable NSString *)releaseName
                     environment:(nullable NSString *)environment
-                    transaction:(nullable NSString *)transaction
-                         userId:(nullable NSString *)userId
                     userSegment:(nullable NSString *)userSegment
+                     sampleRate:(nullable NSString *)sampleRate
 {
 
     if (self = [super init]) {
@@ -24,9 +23,8 @@
         _publicKey = publicKey;
         _releaseName = releaseName;
         _environment = environment;
-        _transaction = transaction;
-        _userId = userId;
         _userSegment = userSegment;
+        _sampleRate = sampleRate;
     }
 
     return self;
@@ -35,7 +33,7 @@
 - (NSString *)toHTTPHeader
 {
     NSMutableDictionary *information =
-        @{ @"sentry-traceid" : _traceId.sentryIdString, @"sentry-publickey" : _publicKey }
+        @{ @"sentry-trace_id" : _traceId.sentryIdString, @"sentry-public_key" : _publicKey }
             .mutableCopy;
 
     if (_releaseName != nil)
@@ -44,14 +42,11 @@
     if (_environment != nil)
         [information setValue:_environment forKey:@"sentry-environment"];
 
-    if (_transaction != nil)
-        [information setValue:_transaction forKey:@"sentry-transaction"];
-
-    if (_userId != nil)
-        [information setValue:_userId forKey:@"sentry-userid"];
-
     if (_userSegment != nil)
-        [information setValue:_userSegment forKey:@"sentry-usersegment"];
+        [information setValue:_userSegment forKey:@"sentry-user_segment"];
+
+    if (_sampleRate != nil)
+        [information setValue:_sampleRate forKey:@"sentry-sample_rate"];
 
     return [SentrySerialization baggageEncodedDictionary:information];
 }