+++ Tests/SentryTests/SentryHubTests.swift
@@ -1,3 +1,4 @@
+import Sentry
 import XCTest
 
 class SentryHubTests: XCTestCase {
@@ -241,6 +242,28 @@ class SentryHubTests: XCTestCase {
         XCTAssertEqual(tracer.name, fixture.transactionName)
         XCTAssertEqual(customSamplingContext?["customKey"] as? String, "customValue")
         XCTAssertEqual(span.context.operation, fixture.transactionOperation)
+    }
+    
+    func testStartTransaction_checkContextSampleRate_fromOptions() {
+        let options = fixture.options
+        options.tracesSampleRate = 0.49
+        
+        let span = fixture.getSut().startTransaction(transactionContext: TransactionContext(name: fixture.transactionName, operation: fixture.transactionOperation), customSamplingContext: ["customKey": "customValue"])
+        let context = span.context as? TransactionContext
+        
+        XCTAssertEqual(context?.sampleRate, 0.49)
+    }
+    
+    func testStartTransaction_checkContextSampleRate_fromSampler() {
+        let options = fixture.options
+        options.tracesSampler = {  _ -> NSNumber in
+            return NSNumber(value: 0.51)
+        }
+        
+        let span = fixture.getSut().startTransaction(transactionContext: TransactionContext(name: fixture.transactionName, operation: fixture.transactionOperation), customSamplingContext: ["customKey": "customValue"])
+        let context = span.context as? TransactionContext
+        
+        XCTAssertEqual(context?.sampleRate, 0.51)
     }
     
     func testStartTransactionNotSamplingUsingSampleRate() {