+++ Tests/SentryTests/Transaction/SentryBaggageTests.swift
@@ -5,15 +5,14 @@ import XCTest
 class SentryBaggageTests: XCTestCase {
     
     func test_baggageToHeader() {
-        let header = SentryBaggage(trace: SentryId.empty, publicKey: "publicKey", releaseName: "release name", environment: "teste", transaction: "some transaction", userId: "user", userSegment: "test user").toHTTPHeader()
+        let header = SentryBaggage(trace: SentryId.empty, publicKey: "publicKey", releaseName: "release name", environment: "teste", userSegment: "test user", sampleRate: "0.49").toHTTPHeader()
         
-        XCTAssertEqual(header, "sentry-environment=teste,sentry-publickey=publicKey,sentry-release=release%20name,sentry-traceid=00000000000000000000000000000000,sentry-transaction=some%20transaction,sentry-userid=user,sentry-usersegment=test%20user")
+        XCTAssertEqual(header, "sentry-environment=teste,sentry-public_key=publicKey,sentry-release=release%20name,sentry-sample_rate=0.49,sentry-trace_id=00000000000000000000000000000000,sentry-user_segment=test%20user")
     }
     
     func test_baggageToHeader_onlyTrace_ignoreNils() {
-        let header = SentryBaggage(trace: SentryId.empty, publicKey: "publicKey", releaseName: nil, environment: nil, transaction: nil, userId: nil, userSegment: nil).toHTTPHeader()
+        let header = SentryBaggage(trace: SentryId.empty, publicKey: "publicKey", releaseName: nil, environment: nil, userSegment: nil, sampleRate: nil).toHTTPHeader()
         
-        XCTAssertEqual(header, "sentry-publickey=publicKey,sentry-traceid=00000000000000000000000000000000")
+        XCTAssertEqual(header, "sentry-public_key=publicKey,sentry-trace_id=00000000000000000000000000000000")
     }
-    
 }